+++ src/rules/mod.rs
@@ -1,4 +1,4 @@
-pub(crate) mod key_and_en_matches;
+pub(crate) mod key_and_eng_matches;
 pub(crate) mod missing_translations;
 
 use crate::LocalizedTexts;
